% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_graph.R
\name{bar_graph}
\alias{bar_graph}
\title{Creates a customizable PRISM-like bar graph. Currently, error bars and significant bars only work if you want ggplot to graph a stat summary.}
\usage{
bar_graph(
  df,
  x_variable = "",
  y_variable = "",
  group_order = c(),
  group_label_size = 5,
  statistics = mean,
  control = NULL,
  bar_width = 0.4,
  title = NULL,
  title_position = NULL,
  x_label = NULL,
  y_label = NULL,
  x_label_angle = 0,
  minor_ticks = TRUE,
  bar_colors = c(),
  individual_points = TRUE,
  point_size = 3,
  point_shapes = c(),
  point_colors = c(),
  error_bars = TRUE,
  error_bar_width = 0.2,
  error_bar_method = mean_se,
  signif_bars = TRUE,
  comparision_groups = c(),
  signif_y_position = NULL,
  signif_step_increase = 0.2,
  margins = NULL,
  y_axis_limits = NULL
)
}
\arguments{
\item{df}{The data frame placed into ggplot.}

\item{x_variable}{The independent variable. The bar labels. Written as a string name of a column.}

\item{y_variable}{The dependent variable. The measured values. Written as a string name of a column}

\item{group_label_size}{The font size of the x variable labels. Default is 5.}

\item{statistics}{A string or function. The statistics placed in stat_summary. Default is mean.}

\item{control}{A string. The x variable that is the control of the df. Is used in fold change calculations.}

\item{bar_width}{Integer or float. The width of the bars. Default is 0.4.}

\item{title}{The title of the graph. Written as a string.}

\item{title_position}{Edits the margins around the title. Enter as margin\(top, right, bottom, left, unit\). Example: margin\(0,0,0,0, "cm"\).}

\item{x_label}{Adds a label under the x-axis. Written as a string.}

\item{y_label}{Adds a label next to the y-axis. Written as a string.}

\item{x_label_angle}{Integer or float. Angle of the bar group labels. Default is 0.}

\item{minor_ticks}{Boolean value. Adds minor ticks.}

\item{bar_colors}{Changes the bar colors. Write as a vector with hex color strings.}

\item{individual_points}{Boolean value. Adds jitter points.}

\item{point_size}{Integer or float. Changes the size of the individual points. Default is 3.}

\item{point_shapes}{A numeric vector. Changes the shape of the individual points.}

\item{error_bars}{Boolean value. Adds error bars to the graph.}

\item{error_bar_width}{Integer or float. Changes the width of the error bars. Default is 0.2.}

\item{error_bar_method}{A string or function. What the error bars plot. Default is mean of standard error.}

\item{signif_bars}{Boolean value. Adds significant bars.}

\item{comparision_groups}{A list argument of what vector pairs to conduct a t-tests on. Example: list\(c\(x1, x2\), c\(x2, x3\)\).}

\item{signif_y_position}{A numeric vector. Changes the specific y-position of the significant bars.}

\item{signif_step_increase}{A numeric vector.Minimizes overlap by increasing in fraction of total height for every bar.}

\item{margins}{Edits the margins around the graph. Enter as margin\(top, right, bottom, left, unit\). Example: margin\(0,0,0,0, "cm"\).}

\item{y_axis_limits}{Changes the limits of the y-axis. Write as c(min, max).}

\item{point_color}{Changes the color of the individual points. Write as a vector with hex color strings.}
}
\value{
A PRISM-themed ggplot graph.
}
\description{
Creates a customizable PRISM-like bar graph. Currently, error bars and significant bars only work if you want ggplot to graph a stat summary.
}
