% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold_change.R
\name{fold_change}
\alias{fold_change}
\title{A fold_change closure function to use in stat summary}
\usage{
fold_change(x)
}
\arguments{
\item{x}{The groups}
}
\description{
A fold_change closure function to use in stat summary
}
