% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rout_outlier.R
\name{rout_outlier}
\alias{rout_outlier}
\title{Does the ROUT outlier test defined in PRISM.}
\usage{
rout_outlier(df, x, y, Q = 0.01, outliers = FALSE)
}
\arguments{
\item{df}{The data frame where outliers need to be removed}

\item{x}{A string. The column which contains the groups you want to compare; the independent variable.}

\item{y}{A string. The column that contains the measured value; the dependent variable.}

\item{Q}{Controls the false discovery rate \(FDR\). Ex. if Q=0.01, then at least 99% of the identified outliers would be actual outliers. Higher Q means that the threshold for defining outliers is less strict.}

\item{outliers}{Returns the whole dataframe with outliers defined if TRUE. Removes the outliers from the dataframe if FALSE.}
}
\description{
Source: Motulsky HM and Brown RE, Detecting outliers when fitting data with nonlinear regression – a new method based on robust nonlinear regression and the false discovery rate, BMC Bioinformatics 2006, 7:123.
}
